Refactor: rename package to more suitable
