+++ core/favoritewatchlist/src/main/kotlin/com/waffiq/bazz_movies/core/favoritewatchlist/utils/helpers/SnackBarUserLoginData.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.favoritewatchlist.utils.helpers
 
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.FavoritePostModel
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.WatchlistPostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.FavoritePostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.WatchlistPostModel
 
 data class SnackBarUserLoginData(
   val isSuccess: Boolean,
