+++ core/network/src/main/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/MovieDataSource.kt
@@ -3,11 +3,11 @@ package com.waffiq.bazz_movies.core.network.data.remote.datasource
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.waffiq.bazz_movies.core.network.data.remote.paging_sources.GenericPagingSource
-import com.waffiq.bazz_movies.core.network.data.remote.paging_sources.SearchPagingSource
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.FavoritePostModel
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.RatePostModel
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.WatchlistPostModel
+import com.waffiq.bazz_movies.core.network.data.remote.pagingsources.GenericPagingSource
+import com.waffiq.bazz_movies.core.network.data.remote.pagingsources.SearchPagingSource
+import com.waffiq.bazz_movies.core.network.data.remote.models.FavoritePostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.RatePostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.WatchlistPostModel
 import com.waffiq.bazz_movies.core.network.data.remote.responses.omdb.OMDbDetailsResponse
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.ResultItemResponse
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.StatedResponse
