+++ core/network/src/main/kotlin/com/waffiq/bazz_movies/core/network/data/remote/retrofit/TMDBApiService.kt
@@ -1,8 +1,8 @@
 package com.waffiq.bazz_movies.core.network.data.remote.retrofit
 
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.FavoritePostModel
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.RatePostModel
-import com.waffiq.bazz_movies.core.network.data.remote.post_body.WatchlistPostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.FavoritePostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.RatePostModel
+import com.waffiq.bazz_movies.core.network.data.remote.models.WatchlistPostModel
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.MovieTvResponse
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.StatedResponse
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.account.AccountDetailsResponse
